+++ Frontend/PineappleSite.Presentation/Models/Products/UpdateProductViewModel.cs
@@ -22,7 +22,7 @@ public class UpdateProductViewModel
     
     [Required(ErrorMessage = "Поле обязательно к заполнению")]
     [Range(5, 1000, ErrorMessage = "Стоимость товара доложна быть более 5, но не более 1000")]
-    public decimal Price { get; init; }
+    public double Price { get; init; }
 
     public IFormFile? Avatar { get; init; }
